Revert HttpTileSourceDefinition, merge back into HttpTileSource - we need an immutable tile source definition that can be serialized
