+++ Tests/BruTile.Tests/Web/HttpTileSourceTests.cs
@@ -1,11 +1,12 @@
 ﻿// Copyright (c) BruTile developers team. All rights reserved. See License.txt in the project root for license information.
 
-using System;
 using System.IO;
+using System.Net;
 using System.Net.Http;
+using System.Threading;
 using System.Threading.Tasks;
-using BruTile.Extensions;
 using BruTile.Predefined;
+using BruTile.Web;
 using NUnit.Framework;
 using RichardSzalay.MockHttp;
 
@@ -15,22 +16,59 @@ namespace BruTile.Tests.Web;
 public class HttpTileSourceTests
 {
     [Test]
-    public static async Task TestAsyncTileFetcher()
+    public async Task TestGetTileAsync()
     {
         // Arrange
-        var tileSource = KnownTileSources.Create();
+        var tileInfo = new TileInfo { Index = new TileIndex(3, 5, 7) };
+        var httpTileSource = CreateHttpTileSource();
         var mockHttp = new MockHttpMessageHandler();
-        mockHttp.When("https://*").Respond("image/png", new MemoryStream());
+        mockHttp.Expect(httpTileSource.GetUrl(tileInfo).ToString())
+            .Respond("image/png", new MemoryStream([0x01, 0x02, 0x03, 0x04]));
         var httpClient = new HttpClient(mockHttp);
         httpClient.DefaultRequestHeaders.UserAgent.ParseAdd("BruTile.Tests");
-        var range = tileSource.Schema.GetTileInfos(tileSource.Schema.Extent, 3);
-        var timeStart = DateTime.Now;
 
         // Act
-        var tiles = await httpClient.GetTilesAsync(tileSource, range).ConfigureAwait(false);
+        var response = await httpTileSource.GetTileAsync(httpClient, tileInfo, CancellationToken.None).ConfigureAwait(false);
 
         // Assert
-        Console.WriteLine("Durations: {0:0} milliseconds", DateTime.Now.Subtract(timeStart).TotalMilliseconds);
-        Assert.AreEqual(64, tiles.Length);
+        Assert.NotNull(response);
+        Assert.AreEqual(new byte[] { 0x01, 0x02, 0x03, 0x04 }, response);
+    }
+
+    [TestCase("UserAgentOverride", "UserAgentOverride", "The 'User-Agent' header should be overridden.")]
+    [TestCase(null, "DefaultUserAgent", "The 'User-Agent' header should have the default value.")]
+    public async Task TestUserAgentOverride(string userAgentOverride, string expectedUserAgent, string message)
+    {
+        // Arrange
+        var httpTileSource = CreateHttpTileSource(userAgentOverride);
+        var tileInfo = new TileInfo { Index = new TileIndex(3, 5, 7) };
+        var url = httpTileSource.GetUrl(tileInfo);
+
+        var mockHttp = new MockHttpMessageHandler();
+        _ = mockHttp.When(url.ToString())
+            .Respond(request =>
+            {
+                Assert.IsTrue(request.Headers.UserAgent.ToString().Contains(expectedUserAgent), message); // Check if UserAgent header is correctly set
+                return new HttpResponseMessage(HttpStatusCode.OK);
+            });
+
+        var httpClient = new HttpClient(mockHttp);
+        httpClient.DefaultRequestHeaders.UserAgent.ParseAdd("DefaultUserAgent");
+
+        // Act
+        var response = await httpTileSource.GetTileAsync(httpClient, tileInfo, CancellationToken.None);
+
+        // Assert
+        Assert.NotNull(response);
+    }
+
+    private static HttpTileSource CreateHttpTileSource(string? userAgentOverride = null)
+    {
+        var tileSchema = new GlobalSphericalMercator();
+        var basicUrlBuilder = new BasicUrlBuilder("http://localhost/{z}/{x}/{y}.png");
+        var name = "name";
+        var attribution = new Attribution("attribution");
+        return new HttpTileSource(tileSchema, basicUrlBuilder, name, null, attribution,
+            (userAgentOverride is null) ? null : (m) => m.Headers.UserAgent.ParseAdd(userAgentOverride));
     }
 }
